# openapi.yaml
openapi: 3.0.0
info:
  title: User Authentication API
  version: 1.0.0
  description: |
    Comprehensive API documentation for the User Authentication Service with Resource-Based Access Control (RBAC).
    
    This API provides:
    - User authentication and authorization
    - Global role-based access control
    - Resource-specific role and permission management
    - OpenID Connect (OIDC) configuration
    - JWT token management
    
    ## Authentication
    Most endpoints require authentication via JWT Bearer tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Resource-Based Access Control
    The API supports both global roles (ADMIN, SUPERADMIN) and resource-specific roles.
    - Users can have multiple global roles
    - Users can have different roles for different resources
    - A user can only have one role per resource
    - Permissions are verb-based (read, write, update, delete, etc.)
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user's profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Auth]
      summary: Update current user's profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: User profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request - validation error or old password mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Auth]
      summary: Deactivate current user's account
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Account deactivated successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users:
    get:
      tags: [Users]
      summary: Get all users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
    post:
      tags: [Users]
      summary: Create a new user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request - validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
...
  /roles:
    get:
      tags: [RBAC]
      summary: Get all global roles
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of roles retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
...
  /resources:
    get:
      tags: [Resources]
      summary: Get resources accessible to the authenticated user
      description: Retrieve all resources that the authenticated user has access to based on their assigned roles and permissions. Super admins see all resources, while regular users only see resources they have been granted access to.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of accessible resources retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Resource'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
    post:
      tags: [Resources]
      summary: Create a new resource
      description: Create a new resource and automatically initialize it with an ACTIVE status. A ResourceStatus record is created alongside the resource to track its lifecycle state.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResourceRequest'
      responses:
        '201':
          description: Resource created successfully with ACTIVE status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        '400':
          description: Bad request - validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
...
  /resources/roles:
    post:
      tags: [Resources]
      summary: Create a role for a resource
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResourceRole'
      responses:
        '201':
          description: Role created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceRole'
        '400':
          description: Bad request - validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
...
  /config/permissions:
    get:
      tags: [Config]
      summary: Get role-permission mapping
      description: Retrieve the system-wide default role-to-permission mapping
      responses:
        '200':
          description: Role-permission mapping retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    role:
                      type: string
                      example: "ADMIN"
                    permissions:
                      type: array
                      items:
                        type: string
                        example: "read:users"
        '500':
          description: Internal server error

  # OpenID Connect Endpoints
...
  /user-details:
    get:
      tags: [UserDetails]
      summary: Get all user details
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of user details retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /permissions:
    get:
      tags: [Permissions]
      summary: Get all available permissions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of permissions retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - insufficient permissions
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization: Bearer {token}"'
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
        email:
          type: string
          format: email
          description: User email address
        isActive:
          type: boolean
          description: Whether the user account is active
        isSuperAdmin:
          type: boolean
          description: Whether the user has super admin privileges
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        roles:
          type: array
          items:
            $ref: '#/components/schemas/UserRole'
          description: Global roles assigned to the user
    UserRole:
      type: object
      properties:
        role:
          $ref: '#/components/schemas/Role'
    CreateUser:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          format: password
          minLength: 6
          description: User password (minimum 6 characters)
        isSuperAdmin:
          type: boolean
          description: Whether the user should be a super admin
        firstName:
          type: string
          description: User's first name
        lastName:
          type: string
          description: User's last name
        nickName:
          type: string
          description: User's nickname
        contactNumber:
          type: string
          description: User's contact number
        reportTo:
          type: string
          format: uuid
          description: ID of the user's manager
...
    UserDetails:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID
        firstName:
          type: string
          nullable: true
          description: User's first name
        lastName:
          type: string
          nullable: true
          description: User's last name
        nickName:
          type: string
          nullable: true
          description: User's nickname
        contactNumber:
          type: string
          nullable: true
          description: User's contact number
        reportToId:
          type: string
          format: uuid
          nullable: true
          description: ID of the user's manager
        createdAt:
          type: string
          format: date-time
          description: Details creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        user:
          type: object
          properties:
            email:
              type: string
              format: email
              description: User email
            isSuperAdmin:
              type: boolean
              description: Whether user is super admin
            isActive:
              type: boolean
              description: Whether user is active
        reportTo:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
              description: Manager's user ID
            email:
              type: string
              format: email
              description: Manager's email
            details:
              type: object
              nullable: true
              properties:
                firstName:
                  type: string
                  nullable: true
                  description: Manager's first name
                lastName:
                  type: string
                  nullable: true
                  description: Manager's last name
                nickName:
                  type: string
                  nullable: true
                  description: Manager's nickname
...
    Role:
      type: object
      properties:
        id:
          type: string
          description: Role identifier
        name:
          type: string
          description: Role name
          example: ADMIN
        description:
          type: string
          nullable: true
          description: Role description
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/RolePermission'
          description: Permissions associated with this role
...
    Permission:
      type: object
      properties:
        id:
          type: string
          description: Permission identifier
        name:
          type: string
          description: Permission name
          example: read_users
        description:
          type: string
          nullable: true
          description: Permission description
    RolePermission:
      type: object
      properties:
        permission:
          $ref: '#/components/schemas/Permission'

    # Resource-based RBAC schemas
    Resource:
      type: object
      properties:
        id:
          type: string
          description: Resource identifier
        name:
          type: string
          description: Resource name
        description:
          type: string
          nullable: true
          description: Resource description
        createdAt:
          type: string
          format: date-time
          description: Resource creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
...
    CreateResourceRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Resource name
        description:
          type: string
          description: Optional resource description
        id:
          type: string
          description: Optional custom resource ID. If not provided, an auto-generated ID will be used.
      description: Request payload for creating a new resource. Upon successful creation, a ResourceStatus record will be automatically created with ACTIVE status.
    ResourceRole:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Resource role identifier
        name:
          type: string
          description: Role name within the resource context
          example: editor
        resourceType:
          type: string
          description: Type of the resource
          example: Organization
        resourceId:
          type: string
          description: Unique identifier of the resource
          example: org_123
        description:
          type: string
          nullable: true
          description: Role description
        permissions:
          type: array
          items:
            type: string
          description: Permission verbs associated with this resource role
          example: ['read', 'update']
        createdAt:
          type: string
          format: date-time
          description: Role creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    CreateResourceRole:
      type: object
      required:
        - name
        - permissions
      properties:
        name:
          type: string
          description: Role name
          example: editor
        description:
          type: string
          description: Optional role description
        resourceId:
          type: string
          description: Optional resource ID for resource-specific roles
        permissions:
          type: array
          items:
            type: string
          minItems: 1
          description: Array of permission verbs
          example: ['read', 'update']
...
    UpdateProfileRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: New email address
        oldPassword:
          type: string
          format: password
          description: Current password (required if changing password)
        newPassword:
          type: string
          format: password
          minLength: 6
          description: New password (minimum 6 characters)
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
...
security:
  - bearerAuth: []